Untitled 2D Game - Walk Test v1 WEB RELEASE
The first version of the game, without a title as of now. The player can walk and jump with animations, the player also has to touch the floor before jumping again once they reach a certain height or if they let go of the space key mid-jump. Player designed by Yemm and programming was done by Me, LogEdude!